[BEAM-444] Promote isBlockOnRun to PipelineOptions
